おゆずりコメント部分を切り出した
